requirements.txt
boto3==1.9.234  # pyup: ignore
django-classified==0.9
django-environ==0.4.5
django-storages==1.9.1
django==3.0.6
gunicorn==20.0.4
psycopg2-binary==2.8.5
python-memcached==1.59
whitenoise==5.0.1
django-allauth==0.41.0
django-bootstrap4==2.3.1
sorl-thumbnail==12.6.3
django-storages==1.9.1
django-fontawesome-5==1.0.18

